+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Hide reposts from public timelines.
+- Validate media descriptions before saving to database.
 
 ## [2.6.0] - 2023-12-25